requirements.txt
Flask==3.0.0
requests==2.31.0
flask-dance[sqla]
python-dotenv
cryptography
flask-sqlalchemy
